requirements.txt
# ========== Web Framework & REST API ==========
Django==6.0.3
djangorestframework==3.16.1
gunicorn==25.1.0

# ========== Data Processing & ML Core ==========
pandas==3.0.1
numpy==1.22.0
scikit-learn==1.3.2
joblib==1.4.2

# ========== Deep Learning ==========
TensorFlow==2.12.0
Keras==3.13.2

# ========== Visualization & Charts ==========
plotly==6.6.0
matplotlib==3.7.5

# ========== HTTP & Utilities ==========
requests==2.31.0

# ========== Configuration & Environment ==========
python-dotenv==1.2.2

# ========== Development & Production Tools ==========
wheel==0.42.0
setuptools==70.0.0

